requirements.txt
# Requirements for the project
pypdf >= 4.0.0
python-docx >= 1.2.0
beautifulsoup4 >= 4.14.3
requests >= 2.32.5
pytest >= 9.0.2
sentence-transformers >= 5.2.2
faiss-cpu >= 1.13.2
pydantic >= 2.12.5
accelerate >= 1.12.0
fastapi[standard] >= 0.128.3
